+++ src/views/CreateOrEdit.vue
@@ -1,9 +1,7 @@
 <script lang="ts">
-import { computed, defineComponent, ref, onMounted } from 'vue'
-import { useRouter } from 'vue-router'
-import * as Services from '../services/'
+import { defineComponent } from 'vue'
 import FormTask from '../components/FormTask.vue'
-import type { Task } from '@/types'
+import useCreateOrEdit from '../composables/useCreateOrEdit'
 
 export default defineComponent({
   props: {
@@ -13,54 +11,13 @@ export default defineComponent({
     FormTask
   },
   setup(props) {
-    const router = useRouter()
-
-    const task = ref({} as Task)
-
-    const pending = ref(false)
-
-    const isRenderable = computed(
-      ()=> (props.id && Object.keys(task.value).length > 0)
-        || props.id===undefined
-    )
+    const {
+      isRenderable,
+      pending,
+      task,
     
-    const getTask = ()=> {
-      pending.value = true
-      Services.getTask(props.id)
-        .then(response => task.value = response.data)
-        .catch(
-          error => console.log({
-            errorCode: error.code, errorMessage: error.message
-          })
-        )
-        .finally(() => pending.value = false)
-    }
-
-    const submit = (payload: Task) => {
-      pending.value = true
-      if (props.id===undefined) {
-        Services.insertTask(payload)
-          .then(response => {
-            alert(response.data.message)
-            router.push({name: 'index'})
-          })
-          .catch(error => console.log(error))
-          .finally(() => pending.value = false)
-      } else {      
-        Services.updateTask(props.id, payload)
-          .then(response => {
-            alert(response.data.message)
-            router.push({name: 'index'})
-          })
-          .catch(error => console.log(error))
-          .finally(() => pending.value = false)
-      }
-    }
-
-    onMounted(()=>{
-      if (props.id)
-        getTask();
-    })
+      submit
+    } = useCreateOrEdit(props)
 
     return {
       isRenderable,
